Create IndirectHeap
Changes made to original heap to make a indirectheap adding an array for into/outof and changing original array into a key. The into/outof arrays sync with each other but the ordering is currently incorrect. Will need to work on Siftdown which I think is where the problem is.